Rename everr-resource-usage action to everr-action
Move the action to .github/actions/everr-action/, fold into the pnpm
workspace as @everr/action, and update all 11 in-repo call sites.
Drops the standalone npm package-lock in favor of the root pnpm lockfile.
No behavior change; dist regenerated under pnpm.
