--- .github/actions/everr-resource-usage/action.yml
@@ -1,13 +1,0 @@
-name: Everr Resource Usage
-description: Collect per-job machine resource usage and upload a best-effort artifact
-
-inputs:
-  check-run-id:
-    description: Check run id for the current workflow job
-    required: true
-
-runs:
-  using: node24
-  main: dist/index.js
-  post: dist/index.js
-  post-if: always()
