--- .github/actions/everr-resource-usage/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "lib": ["ES2022"],
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext",
-    "noEmitOnError": true,
-    "skipLibCheck": true,
-    "strict": true,
-    "target": "ES2022",
-    "types": ["node"],
-    "verbatimModuleSyntax": true
-  },
-  "include": ["src/index.ts", "scripts/finalize.ts"]
-}
