--- .github/actions/everr-resource-usage/tsconfig.typecheck.json
@@ -1,8 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "allowImportingTsExtensions": true,
-    "noEmit": true
-  },
-  "include": ["src/**/*.ts", "scripts/**/*.ts"]
-}
