+++ .github/workflows/build-and-push-images.yml
@@ -73,7 +73,7 @@ jobs:
         uses: actions/checkout@v6
 
       - name: Collect resource usage
-        uses: ./.github/actions/everr-resource-usage
+        uses: ./.github/actions/everr-action
         with:
           check-run-id: ${{ job.check_run_id }}
 
@@ -108,7 +108,7 @@ jobs:
         uses: actions/checkout@v6
 
       - name: Collect resource usage
-        uses: ./.github/actions/everr-resource-usage
+        uses: ./.github/actions/everr-action
         with:
           check-run-id: ${{ job.check_run_id }}
 
@@ -143,7 +143,7 @@ jobs:
         uses: actions/checkout@v6
 
       - name: Collect resource usage
-        uses: ./.github/actions/everr-resource-usage
+        uses: ./.github/actions/everr-action
         with:
           check-run-id: ${{ job.check_run_id }}
