+++ .github/workflows/build-and-test-collector.yml
@@ -33,7 +33,7 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Collect resource usage
-        uses: ./.github/actions/everr-resource-usage
+        uses: ./.github/actions/everr-action
         with:
           check-run-id: ${{ job.check_run_id }}
 
@@ -73,7 +73,7 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Collect resource usage
-        uses: ./.github/actions/everr-resource-usage
+        uses: ./.github/actions/everr-action
         with:
           check-run-id: ${{ job.check_run_id }}
 
@@ -132,7 +132,7 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Collect resource usage
-        uses: ./.github/actions/everr-resource-usage
+        uses: ./.github/actions/everr-action
         with:
           check-run-id: ${{ job.check_run_id }}
 
@@ -171,7 +171,7 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Collect resource usage
-        uses: ./.github/actions/everr-resource-usage
+        uses: ./.github/actions/everr-action
         with:
           check-run-id: ${{ job.check_run_id }}
