+++ .github/workflows/build-and-test-desktop-app.yml
@@ -46,7 +46,7 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Collect resource usage
-        uses: ./.github/actions/everr-resource-usage
+        uses: ./.github/actions/everr-action
         with:
           check-run-id: ${{ job.check_run_id }}
