+++ .github/workflows/build-everr-cli.yml
@@ -39,7 +39,7 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Collect resource usage
-        uses: ./.github/actions/everr-resource-usage
+        uses: ./.github/actions/everr-action
         with:
           check-run-id: ${{ job.check_run_id }}
