+++ .github/workflows/deploy-desktop-app.yml
@@ -25,7 +25,7 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Collect resource usage
-        uses: ./.github/actions/everr-resource-usage
+        uses: ./.github/actions/everr-action
         with:
           check-run-id: ${{ job.check_run_id }}
